Usage within Julia

GhostKnockoffGWAS is a regular Julia package, which can be used directly within Julia for greater flexibility. To install it, execute the following in Julia

using Pkg
Pkg.add(url="https://github.com/biona001/ghostbasil_jll.jl")
Pkg.add(url="https://github.com/biona001/Ghostbasil.jl")
Pkg.add(url="https://github.com/biona001/GhostKnockoffGWAS")

Warning

This package currently only works on Julia 1.8.x and 1.9.x. If you need it to work on a different Julia version, let us know by filing an issue on Github.

Usage example

The following example performs summary-statistics GWAS under the GhostKnockoff framework.

using GhostKnockoffGWAS

# file paths and directories
knockoff_dir = "/home/groups/sabatti/.julia/dev/GhostKnockoffGWAS/data/EUR"
zfile = "/home/groups/sabatti/.julia/dev/GhostKnockoffGWAS/data/AD_Zscores_Meta_modified.txt"
outdir = "/home/groups/sabatti/.julia/dev/GhostKnockoffGWAS/data"
outfile = "test_alzheimers_meta"

# specify sample size and human genome build
N = 506200
hg_build = 38

# read Z-scores using built-in function read_zscores
z, chr, pos, effect_allele, non_effect_allele = GhostKnockoffGWAS.read_zscores(zfile)

# run analysis
@time ghostbasil_parallel(knockoff_dir, z, chr, pos, effect_allele, 
    non_effect_allele, N, hg_build, outdir, outname=outfile, 
    seed=2023)

Function API

GhostKnockoffGWAS.MK_statistics — Method

MK_statistics(T0::Vector, Tk::Vector{Vector}; filter_method)

Computes the multiple knockoff statistics kappa, tau, and W.

Inputs

T0: p-vector of importance score for original variables

Tk: Vector storing T1, ..., Tm, where Ti is importance scores for the ith knockoff copy

filter_method: Either Statistics.median (default) or max (original function used in 2019 Gimenez and Zou)

output

κ: Index of the most significant feature (κ[i] = 0 if original feature most important, otherwise κ[i] = k if the kth knockoff is most important)

τ: τ[i] stores the most significant statistic among original and knockoff variables minus filter_method() applied to the remaining statistics.

W: coefficient difference statistic W[i] = abs(T0[i]) - abs(Tk[i])

Source

https://github.com/biona001/Knockoffs.jl/blob/master/src/threshold.jl#L96

GhostKnockoffGWAS.find_matching_indices — Method

find_matching_indices(a, b)

Returns output vector c such that c[i] contains all indices of b that matches a[i]. Generalizes indexin in Base.

Example

a = [1, 2, 3, 2, 6]
b = [2, 1, 2, 4, 3]
c = find_matching_indices(a, b)
println(c)  # Output: [[2], [1, 3], [5], [1, 3], Int[]]

GhostKnockoffGWAS.ghost_knockoffs — Method

ghost_knockoffs(Zscores, D, Σinv; [m=1])

Generate Ghost knockoffs given a list of z-scores (GWAS summary statistic).

Inputs

Zscores: List of z-score statistics

D: Matrix obtained from solving the knockoff problem satisfying (m+1)/m*Σ - D ⪰ 0

Σinv: Inverse of the covariance matrix

optional inputs

m: Number of knockoffs

Reference

He, Z., Liu, L., Belloy, M. E., Le Guen, Y., Sossin, A., Liu, X., ... & Ionita-Laza, I. (2021). Summary statistics knockoff inference empowers identification of putative causal variants in genome-wide association studies.

Source

https://github.com/biona001/Knockoffs.jl/blob/master/src/ghost.jl#L32

GhostKnockoffGWAS.mk_threshold — Method

mk_threshold(τ::Vector{T}, κ::Vector{Int}, m::Int, q::Number)

Chooses the multiple knockoff threshold τ̂ > 0 by setting τ̂ = min{ t > 0 : (1/m + 1/m * {#j: κ[j] ≥ 1 and W[j] ≥ t}) / {#j: κ[j] == 0 and W[j] ≥ τ̂} ≤ q }.

Inputs

τ: τ[i] stores the feature importance score for the ith feature, i.e. the value T0 - median(T1,...,Tm). Note in Gimenez and Zou, the max function is used instead of median

κ: κ[i] stores which of m knockoffs has largest importance score. When original variable has largest score, κ[i] == 0.

m: Number of knockoffs per variable generated

q: target FDR (between 0 and 1)

rej_bounds: Number of values of top τ to consider (default = 10000)

Reference:

Equations 8 and 9 in supplement of "Identification of putative causal loci in wholegenome sequencing data via knockoff statistics" by He et al.

Algorithm 1 of "Improving the Stability of the Knockoff Procedure: Multiple Simultaneous Knockoffs and Entropy Maximization" by Gimenez and Zou.

Source

https://github.com/biona001/Knockoffs.jl/blob/master/src/threshold.jl#L55

GhostKnockoffGWAS.read_zscores — Method

read_zscores(filepath::String)

Helper function to read a Z-score file. First row must be a header column with CHR, POS, REF, ALT, and Z. All other columns will be ignored.

todo: detect duplicate SNPs

GhostKnockoffGWAS.sample_mvn_efficient — Method

sample_mvn_efficient(C::AbstractMatrix{T}, D::AbstractMatrix{T}, m::Int)

Efficiently samples from N(0, A) where

\[\begin{aligned} A &= \begin{pmatrix} C & C-D & \cdots & C-D\\ C-D & C & \cdots & C-D\\ \vdots & & \ddots & \vdots\\ C-D & C-D & & C \end{pmatrix} \end{aligned}\]

Note there are m blocks per row/col

Source

https://github.com/biona001/Knockoffs.jl/blob/master/src/ghost.jl#L60